requirements.txt
# Flask and Server
Flask==2.3.3
Flask-CORS==4.0.0
gunicorn==2.1.2.0
python-dotenv==1.0.0
Werkzeug==2.3.7

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Flask-Migrate==4.0.5
mysqlclient==2.2.4

# AWS
boto3==1.33.6

# Task Queue
celery==5.3.6
redis==5.0.1

# ML & Data Processing
opencv-python-headless==4.8.1.78
numpy==1.24.3
pandas==2.2.2
Pillow==10.0.1
torch==2.2.0
transformers==4.39.3
scikit-learn==1.3.0
google-generativeai==0.1.0
path==16.7.1
nltk==3.8.1
onnxruntime==1.16.3
